+++ pyo3-ffi/build.rs
@@ -70,7 +70,7 @@ fn emit_link_config(interpreter_config: &InterpreterConfig) -> Result<()> {
     }
 
     // serialize the whole interpreter config in DEP_PYTHON_PYO3_CONFIG
-    interpreter_config.to_cargo_link_env()?;
+    interpreter_config.to_cargo_dep_env()?;
 
     Ok(())
 }